\name{minimal-package}
\alias{minimal-package}
\alias{minimal}
\docType{package}
\title{\packageTitle{minimal}}
\description{\packageDescription{minimal}}
\section{Package Content}{\packageIndices{minimal}}
\author{\packageAuthor{minimal}}
\section{Maintainer}{\packageMaintainer{minimal}}
\keyword{package}
